third_party/blink/web_tests/external/wpt/css/css-syntax/urange-parsing-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] "u/**/+/**/a/**/?" => "U+A0-AF"
  assert_equals: expected "U+A0-AF" but got "U+1357"
[FAIL] "u/**/+0a/**/?" => "U+A0-AF"
  assert_equals: expected "U+A0-AF" but got "U+1357"
[FAIL] "u/**/+0/**/?" => "U+0-F"
  assert_equals: expected "U+0-F" but got "U+1357"
[FAIL] "u/**/+0/**/-0a" => "U+0-A"
  assert_equals: expected "U+0-A" but got "U+1357"
[FAIL] "u/**/+0/**/-1" => "U+0-1"
  assert_equals: expected "U+0-1" but got "U+1357"
[FAIL] "u/**/+/**/?" => "U+0-F"
  assert_equals: expected "U+0-F" but got "U+1357"
Harness: the test ran to completion.
